internal/featuretests: convert e2e root namespace tests to featuretests
