+++ internal/e2e/lds_test.go
@@ -937,124 +937,6 @@ func TestLDSCustomAccessLogPaths(t *testing.T) {
 	}, streamLDS(t, cc))
 }
 
-func TestLDSIngressRouteInsideRootNamespaces(t *testing.T) {
-	rh, cc, done := setup(t, func(reh *contour.EventHandler) {
-		reh.Builder.Source.RootNamespaces = []string{"roots"}
-	})
-	defer done()
-
-	// assert that there is only a static listener
-	assert.Equal(t, &v2.DiscoveryResponse{
-		VersionInfo: "0",
-		Resources: resources(t,
-			staticListener(),
-		),
-		TypeUrl: listenerType,
-		Nonce:   "0",
-	}, streamLDS(t, cc))
-
-	// ir1 is an ingressroute that is in the root namespace
-	ir1 := &ingressroutev1.IngressRoute{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "simple",
-			Namespace: "roots",
-		},
-		Spec: ingressroutev1.IngressRouteSpec{
-			VirtualHost: &projcontour.VirtualHost{Fqdn: "example.com"},
-			Routes: []ingressroutev1.Route{{
-				Match: "/",
-				Services: []ingressroutev1.Service{{
-					Name: "kuard",
-					Port: 8080,
-				}},
-			}},
-		},
-	}
-
-	svc1 := &v1.Service{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "kuard",
-			Namespace: "roots",
-		},
-		Spec: v1.ServiceSpec{
-			Ports: []v1.ServicePort{{
-				Name:     "http",
-				Protocol: "TCP",
-				Port:     8080,
-			}},
-		},
-	}
-
-	// add ingressroute & service
-	rh.OnAdd(svc1)
-	rh.OnAdd(ir1)
-
-	// assert there is an active listener
-	assert.Equal(t, &v2.DiscoveryResponse{
-		VersionInfo: "1",
-		Resources: resources(t,
-			&v2.Listener{
-				Name:    "ingress_http",
-				Address: envoy.SocketAddress("0.0.0.0", 8080),
-				FilterChains: envoy.FilterChains(
-					envoy.HTTPConnectionManager("ingress_http", envoy.FileAccessLogEnvoy("/dev/stdout"), 0),
-				),
-			},
-			staticListener(),
-		),
-		TypeUrl: listenerType,
-		Nonce:   "1",
-	}, streamLDS(t, cc))
-}
-
-func TestLDSIngressRouteOutsideRootNamespaces(t *testing.T) {
-	rh, cc, done := setup(t, func(reh *contour.EventHandler) {
-		reh.Builder.Source.RootNamespaces = []string{"roots"}
-	})
-	defer done()
-
-	// assert that there is only a static listener
-	assert.Equal(t, &v2.DiscoveryResponse{
-		VersionInfo: "0",
-		Resources: resources(t,
-			staticListener(),
-		),
-		TypeUrl: listenerType,
-		Nonce:   "0",
-	}, streamLDS(t, cc))
-
-	// ir1 is an ingressroute that is not in the root namespaces
-	ir1 := &ingressroutev1.IngressRoute{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "simple",
-			Namespace: "default",
-		},
-		Spec: ingressroutev1.IngressRouteSpec{
-			VirtualHost: &projcontour.VirtualHost{Fqdn: "example.com"},
-			Routes: []ingressroutev1.Route{{
-				Match: "/",
-				Services: []ingressroutev1.Service{{
-					Name: "kuard",
-					Port: 8080,
-				}},
-			}},
-		},
-	}
-
-	// add ingressroute
-	rh.OnAdd(ir1)
-
-	// assert that there is only a static listener
-	assert.Equal(t, &v2.DiscoveryResponse{
-		VersionInfo: "1",
-		Resources: resources(t,
-			staticListener(),
-		),
-		TypeUrl: listenerType,
-		Nonce:   "1",
-	}, streamLDS(t, cc))
-}
-
 func TestIngressRouteHTTPS(t *testing.T) {
 	rh, cc, done := setup(t)
 	defer done()
