+++ internal/e2e/rds_test.go
@@ -1119,115 +1119,6 @@ func TestDefaultBackendDoesNotOverwriteNamedHost(t *testing.T) {
 	}, streamRDS(t, cc, "ingress_http"))
 }
 
-func TestRDSIngressRouteInsideRootNamespaces(t *testing.T) {
-	rh, cc, done := setup(t, func(reh *contour.EventHandler) {
-		reh.Builder.Source.RootNamespaces = []string{"roots"}
-	})
-	defer done()
-
-	rh.OnAdd(&v1.Service{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "kuard",
-			Namespace: "roots",
-		},
-		Spec: v1.ServiceSpec{
-			Ports: []v1.ServicePort{{
-				Protocol:   "TCP",
-				Port:       8080,
-				TargetPort: intstr.FromInt(8080),
-			}},
-		},
-	})
-
-	// ir1 is an ingressroute that is in the root namespaces
-	ir1 := &ingressroutev1.IngressRoute{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "simple",
-			Namespace: "roots",
-		},
-		Spec: ingressroutev1.IngressRouteSpec{
-			VirtualHost: &projcontour.VirtualHost{Fqdn: "example.com"},
-			Routes: []ingressroutev1.Route{{
-				Match: "/",
-				Services: []ingressroutev1.Service{{
-					Name: "kuard",
-					Port: 8080,
-				}},
-			}},
-		},
-	}
-
-	// add ingressroute
-	rh.OnAdd(ir1)
-
-	assert.Equal(t, &v2.DiscoveryResponse{
-		VersionInfo: "1",
-		Resources: resources(t,
-			envoy.RouteConfiguration("ingress_http",
-				envoy.VirtualHost("example.com",
-					&envoy_api_v2_route.Route{
-						Match:  routePrefix("/"),
-						Action: routecluster("roots/kuard/8080/da39a3ee5e"),
-					},
-				),
-			),
-		),
-		TypeUrl: routeType,
-		Nonce:   "1",
-	}, streamRDS(t, cc, "ingress_http"))
-}
-
-func TestRDSIngressRouteOutsideRootNamespaces(t *testing.T) {
-	rh, cc, done := setup(t, func(reh *contour.EventHandler) {
-		reh.Builder.Source.RootNamespaces = []string{"roots"}
-	})
-	defer done()
-
-	rh.OnAdd(&v1.Service{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "kuard",
-			Namespace: "default",
-		},
-		Spec: v1.ServiceSpec{
-			Ports: []v1.ServicePort{{
-				Protocol:   "TCP",
-				Port:       8080,
-				TargetPort: intstr.FromInt(8080),
-			}},
-		},
-	})
-
-	// ir1 is an ingressroute that is not in the root namespaces
-	ir1 := &ingressroutev1.IngressRoute{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "simple",
-			Namespace: "default",
-		},
-		Spec: ingressroutev1.IngressRouteSpec{
-			VirtualHost: &projcontour.VirtualHost{Fqdn: "example.com"},
-			Routes: []ingressroutev1.Route{{
-				Match: "/",
-				Services: []ingressroutev1.Service{{
-					Name: "kuard",
-					Port: 8080,
-				}},
-			}},
-		},
-	}
-
-	// add ingressroute
-	rh.OnAdd(ir1)
-
-	assert.Equal(t, &v2.DiscoveryResponse{
-		VersionInfo: "1",
-		Resources: resources(t,
-			envoy.RouteConfiguration("ingress_http"),
-		),
-		TypeUrl: routeType,
-		Nonce:   "1",
-	}, streamRDS(t, cc, "ingress_http"))
-}
-
 // Test DAGAdapter.IngressClass setting works, this could be done
 // in LDS or RDS, or even CDS, but this test mirrors the place it's
 // tested in internal/contour/route_test.go
@@ -2182,56 +2073,6 @@ func service(ns, name string, ports ...v1.ServicePort) *v1.Service {
 			Ports: ports,
 		},
 	}
-}
-
-func TestRDSHTTPProxyOutsideRootNamespaces(t *testing.T) {
-	rh, cc, done := setup(t, func(reh *contour.EventHandler) {
-		reh.Builder.Source.RootNamespaces = []string{"roots"}
-	})
-	defer done()
-
-	rh.OnAdd(&v1.Service{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "kuard",
-			Namespace: "default",
-		},
-		Spec: v1.ServiceSpec{
-			Ports: []v1.ServicePort{{
-				Protocol:   "TCP",
-				Port:       8080,
-				TargetPort: intstr.FromInt(8080),
-			}},
-		},
-	})
-
-	// proxy1 is an httpproxy that is not in the root namespaces
-	proxy1 := &projcontour.HTTPProxy{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "simple",
-			Namespace: "default",
-		},
-		Spec: projcontour.HTTPProxySpec{
-			VirtualHost: &projcontour.VirtualHost{Fqdn: "example.com"},
-			Routes: []projcontour.Route{{
-				Services: []projcontour.Service{{
-					Name: "kuard",
-					Port: 8080,
-				}},
-			}},
-		},
-	}
-
-	// add proxy
-	rh.OnAdd(proxy1)
-
-	assert.Equal(t, &v2.DiscoveryResponse{
-		VersionInfo: "1",
-		Resources: resources(t,
-			envoy.RouteConfiguration("ingress_http"),
-		),
-		TypeUrl: routeType,
-		Nonce:   "1",
-	}, streamRDS(t, cc, "ingress_http"))
 }
 
 func TestHTTPProxyRouteWithAServiceWeight(t *testing.T) {
